pyproject.toml
[project]
name = "real-time-analytics-database"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faker>=40.15.0",
    "kafka-python>=2.3.1",
]
